Finished & Styled
Add days of the week fully rendered, dynamically. Tested and working with changes to location. Also caching each load, so if you load the same weather for the same location twice (for the same hour) it will rerender. Furthermore, if you let the hour roll over the weather will update for the new current data.